refactor: decompose the model-selection panel into a pull hook and subcomponents
Move the model-download flow (pull + job-event progress/speed/ETA tracking) into
useModelPull, collapsing the 8-line reset block that was copy-pasted across the
success/completed/failed branches into one resetTracking helper. Split the model
button (ModelCard) and the hardware popover (HardwarePopover) into their own files.
